delete dialog confirmation

+++ src/blog/viewArticle.js
@@ -1,13 +1,16 @@
 import {inject} from 'aurelia-framework';
 import {Router} from 'aurelia-router';
 import {ArticleService} from './../services/articleService';
+import {DialogService} from 'aurelia-dialog';
+import {Prompt} from './../common/prompt';
 
-@inject(Router, ArticleService)
+@inject(Router, ArticleService, DialogService)
 
 export class viewArticle{
-    constructor(router, articleService){
+    constructor(router, articleService, dialogService){
         this.router = router;
         this.articleService = articleService;
+        this.dialogService = dialogService;
     }
 
     canActivate(params, routeConfig, $navigationInstruction) {
@@ -15,12 +18,19 @@ export class viewArticle{
             this.article = response[0];
         })
     }
-
+    
     delete(id){
         if(!id) return;
-        this.articleService.deleteArticle(id).then(()=>{
-            this.goToHome();
-        })
+        this.dialogService.open({ viewModel: Prompt, model: 'Good or Bad?'}).then(response => {
+        if (!response.wasCancelled) {
+            this.articleService.deleteArticle(id).then(()=>{
+                this.goToHome();
+            })
+        } else {
+            console.log('bad');
+        }
+        console.log(response.output);
+        });
     }
 
     goToHome(){
